Some warning cleanup.

+++ hw1/miro/PhotonMap.cpp
@@ -14,9 +14,9 @@ void PhotonMap::push_back(const PhotonDeposit& photonDeposit) {
 Vector3 PhotonMap::powerDensity(const Vector3& x, const float& r) {
     Vector3 rho(0, 0, 0);
     if (m_partitionDimensions.length2() <= 0) {
-        for (int i = 0; i < m_photonDeposits.size(); i++) {
+        for (unsigned int i = 0; i < m_photonDeposits.size(); i++) {
             if ((m_photonDeposits[i].m_location - x).length2() < r*r) {
-                rho += m_photonDeposits[i].m_power / (M_PI*r*r);
+                rho += m_photonDeposits[i].m_power / static_cast<float>(M_PI*r*r);
             }
         }
     }
@@ -45,7 +45,7 @@ bool comparePhotons(const std::pair<float,PhotonDeposit>& p1, const std::pair<fl
 
 RadiusDensityPhotons PhotonMap::radiusDensityPhotons(const Vector3& x, const int& n) {
     std::vector<std::pair<float,PhotonDeposit>> displacement2(m_photonDeposits.size());
-    for (int i = 0; i < displacement2.size(); i++) {
+    for (unsigned int i = 0; i < displacement2.size(); i++) {
         displacement2[i] = std::pair<float, PhotonDeposit>((m_photonDeposits[i].m_location - x).length2(), m_photonDeposits[i]);
     }
     std::partial_sort(displacement2.begin(), displacement2.begin() + n, displacement2.end(),comparePhotons);
@@ -55,6 +55,6 @@ RadiusDensityPhotons PhotonMap::radiusDensityPhotons(const Vector3& x, const int
         rdp.m_photons.push_back(displacement2[i].second);
         rdp.m_density += displacement2[i].second.m_power;
     }
-    rdp.m_density /= n;
+    rdp.m_density /= (float) n;
     return rdp;
 }
